+++ src/main/index.ts
@@ -4,15 +4,13 @@ import {network} from '@jovijovi/ether-network';
 import {customConfig} from '../config';
 import {logo} from './logo';
 import {RestfulHandlers} from '../handler';
-import {watchdog} from '../watchdog';
-import * as event from '../event';
+import {module} from '../module';
 
 function main() {
 	log.logo(logo);
 	sys.HandleSignals();
 	config.LoadConfig();
 	customConfig.LoadCustomConfig();
-	http.server.Run(RestfulHandlers);
 
 	network.LoadConfig(customConfig.Get());
 	network.isConnected().then(r => {
@@ -21,9 +19,9 @@ function main() {
 		}
 	});
 
-	watchdog.Run();
-	event.listener.Run();
-	event.fetcher.Run();
+	module.Load();
+
+	http.server.Run(RestfulHandlers);
 }
 
 main();