+++ src/module/id.ts
@@ -1,0 +1,3 @@
+export const ModuleWatchdog = 'watchdog';
+export const ModuleEventListener = 'eventListener';
+export const ModuleEventFetcher = 'eventFetcher';