+++ src/module/index.ts
@@ -1,0 +1,1 @@
+export * as module from './module';